requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.11.0
filelock==3.12.4
fonttools==4.42.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
qrdet==2.3
quadrilateral-fitter==1.8
requests==2.31.0
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.182
urllib3==2.0.4